requirements.txt
annotated-types==0.7.0
Authlib==1.4.1
bandit==1.8.2
boto3==1.36.11
botocore==1.36.11
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dparse==0.6.4
filelock==3.16.1
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jmespath==1.0.1
kafka-python==2.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.2
pbr==6.1.0
pluggy==1.5.0
psutil==6.1.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.19.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.2
safety==3.2.14
safety-schemas==0.0.10
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
stevedore==5.4.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
